small tweaks
I popped in a few commas for clarity in the Introduction paragraph. I'm not sure "in this document" in the second sentence of the definition is clear. What document is this referencing? The profile or the Standards. I would change that to "contains decisions about the  SHOULDS and MAYS for an implementation, as well as...." But if this doesn't read confusing for others, I won't argue the point.

Lois